requirements.txt
nicegui==2.5.0
httpx==0.27.0
pyyaml==6.0.2
python-dotenv==1.0.1
